[actn-interface-tools] Add restconf module. Create empty shell for global-control and yang-utils modules. Add copyrights and delete unessessary comments and unneeded resource files.
restconf module: This module is to send restful request to SDN by using restconf. It acts as a client to send restful requests to the SDN server.
global-control module: This module controls the lifecycle and initiation of objects in other modules.
yang-utils module: This module response for transferring between yang pojo and json.
Also in this commit delete unneeded resource files, add copyrights to the maven pom files and delete some unnessessary comments.

Change-Id: Idfc04649285bd6c5b67b1cc6c2115bf4bb267660
Signed-off-by: ruilin cai <ruilin.cai@huawei.com>
Issue-ID: INT-2213
Signed-off-by: ruilin cai <ruilin.cai@huawei.com>